requirements.txt
openai>=1.12.0
requests==2.31.0
python-dotenv==1.0.0
huggingface-hub==0.20.3
httpx>=0.24.1
numpy>=1.24.3
streamlit>=1.32.0